chore: cuentas

+++ piniabank-backend/Pinia.Application/Interfaces/IServicioCuenta.cs
@@ -4,7 +4,7 @@ namespace Pinia.Application.Interfaces
 {
     public interface IServicioCuenta
     {
-        Task<int> Crear(CrearCuentaDto cuenta);
+        Task<long> Crear(CrearCuentaDto cuenta);
         Task<bool> ActualizarSaldo(int idCuenta, decimal nuevoSaldo);
         Task<bool> CambiarEstado(int idCuenta, bool estado);
         Task<IEnumerable<CuentaDto>> ConsultarPorCliente(int idCliente);
